chore(multisig page): review changes
