+++ frontend/src/app/(routes)/staking/components/ChainUnbondings.tsx
@@ -15,7 +15,7 @@ const ChainUnbondings = ({
   chainName,
 }: ChainUnbondingsProps) => {
   const networks = useAppSelector((state: RootState) => state.wallet.networks);
-  const networkLogo = networks[chainID]?.network.logos.menu;
+  const networkLogo = networks[chainID]?.network?.logos?.menu;
 
   return (
     <>
@@ -26,10 +26,10 @@ const ChainUnbondings = ({
             <UnbondingCard
               key={row.validator_address + entry.completion_time}
               validator={
-                validators?.active[row.validator_address]?.description.moniker
+                validators?.active?.[row.validator_address]?.description.moniker
               }
               identity={
-                validators?.active[row.validator_address]?.description.identity
+                validators?.active?.[row.validator_address]?.description.identity
               }
               chainName={chainName}
               amount={parseDenomAmount(entry.balance, currency.coinDecimals)}
