+++ frontend/src/app/(routes)/staking/components/DialogDelegate.tsx
@@ -10,7 +10,7 @@ import Image from 'next/image';
 import React from 'react';
 import { useForm, Controller } from 'react-hook-form';
 import { dialogBoxStyles } from '../styles';
-import { CLOSE_ICON_PATH } from '@/utils/constants';
+import { CLOSE_ICON_PATH, STAKING_DIALOG_IMAGE_PATH } from '@/utils/constants';
 
 const DialogDelegate = ({
   open,
@@ -70,7 +70,7 @@ const DialogDelegate = ({
           </div>
           <div className="mt-6 mb-[72px] flex gap-6 pr-10 pl-6 items-center">
             <Image
-              src="/delegate-popup-image.png"
+              src={STAKING_DIALOG_IMAGE_PATH}
               height={360}
               width={235}
               alt="Delegate"
