+++ frontend/src/app/(routes)/staking/components/DialogRedelegate.tsx
@@ -11,7 +11,7 @@ import Image from 'next/image';
 import React from 'react';
 import { Controller, useForm } from 'react-hook-form';
 import { dialogBoxStyles } from '../styles';
-import { CLOSE_ICON_PATH } from '@/utils/constants';
+import { CLOSE_ICON_PATH, STAKING_DIALOG_IMAGE_PATH } from '@/utils/constants';
 
 interface ValidatorSet {
   [key: string]: Validator;
@@ -116,7 +116,7 @@ const DialogRedelegate = ({
           </div>
           <div className="mt-6 mb-[72px] flex gap-6 pr-10 pl-6 items-center">
             <Image
-              src="/delegate-popup-image.png"
+              src={STAKING_DIALOG_IMAGE_PATH}
               height={360}
               width={235}
               alt="Redelegate"
