+++ frontend/src/app/(routes)/staking/components/ValidatorLogo.tsx
@@ -1,27 +1,32 @@
 import { ValidatorLogoProps } from '@/types/staking';
 import { VALIDATOR_LOGO_URL } from '@/utils/constants';
-import { ERR_UNKNOWN, GET_VALIDATOR_LOGO_ERROR } from '@/utils/errors';
 import { Avatar } from '@mui/material';
-import Axios, { AxiosError } from 'axios';
+import Axios from 'axios';
 import { get } from 'lodash';
 import React, { useEffect, useState } from 'react';
 
 const ValidatorLogo = ({ identity, width, height }: ValidatorLogoProps) => {
   const [validatorPic, setValidatorPic] = useState<string>('');
+
   useEffect(() => {
+    const CancelToken = Axios.CancelToken;
+    const source = CancelToken.source();
     (async () => {
       try {
-        const { status, data } = await Axios.get(VALIDATOR_LOGO_URL(identity));
+        const { status, data } = await Axios.get(VALIDATOR_LOGO_URL(identity), {
+          cancelToken: source.token,
+        });
 
         if (status === 200)
           setValidatorPic(get(data, 'them[0].pictures.primary.url'));
       } catch (error) {
-        const errorMessage =
-          error instanceof AxiosError ? error.message : ERR_UNKNOWN;
-        console.log(GET_VALIDATOR_LOGO_ERROR, errorMessage);
         setValidatorPic('');
       }
     })();
+
+    return () => {
+      source.cancel();
+    };
   }, [identity]);
   return (
     <>
