+++ frontend/src/utils/constants.ts
@@ -7,3 +7,4 @@ export const VALIDATOR_LOGO_URL = (identity: string) =>
   `https://keybase.io/_/api/1.0/user/lookup.json?key_suffix=${identity}&fields=pictures`;
 export const VALIDATORS_PER_PAGE = 7;
 export const CLOSE_ICON_PATH = '/close-icon.svg';
+export const STAKING_DIALOG_IMAGE_PATH = '/delegate-popup-image.png';
